+++ src/model/compartments/vpm.cpp
@@ -8,8 +8,8 @@ namespace DecoModel {
     //******************************************
     // Override dive segment
     //******************************************
-    void CellVPM::invoke_dive_segment(int time, int start_depth, 
-            int rate, float* gas) {
+    void CellVPM::invoke_dive_segment(uint16_t time, uint16_t start_depth, 
+            uint8_t rate, float* gas) {
 
         dive_segment_buhl(time, start_depth, rate, gas);
         dive_segment_vpm(time, start_depth, rate, gas);
@@ -19,8 +19,8 @@ namespace DecoModel {
     //******************************************
     // Override ceiling wrapper
     //******************************************
-    int CellVPM::get_ceiling() {
-        int ceiling_buhl, ceiling_vpm;
+    uint16_t CellVPM::get_ceiling() {
+        uint16_t ceiling_buhl, ceiling_vpm;
 
         ceiling_buhl = get_ceiling_buhl();
         ceiling_vpm = get_ceiling_vpm();
