+++ src/model/types_constants/types.cpp
@@ -8,8 +8,8 @@ namespace DecoModel {
         //******************************************
         // Assign params for dive segment
         //******************************************
-        Segment::Segment(int seg_start_depth, float seg_rate, 
-                int seg_time, const int* gas_mix)
+        Segment::Segment(uint16_t seg_start_depth, int8_t seg_rate, 
+                uint16_t seg_time, const uint8_t* gas_mix)
                 : start_depth(seg_start_depth), rate(seg_rate), 
                 time(seg_time), gas{gas_mix[0], gas_mix[1]} {}
 
@@ -17,8 +17,8 @@ namespace DecoModel {
         //******************************************
         // Assign params for deco stop
         //******************************************
-        DecoStop::DecoStop(int stop_depth, int stop_time, 
-                const int* gas_mix)
+        DecoStop::DecoStop(uint16_t stop_depth, uint16_t stop_time, 
+                const uint8_t* gas_mix)
                 : depth(stop_depth), time(stop_time), gas{gas_mix[0], gas_mix[1]} {}
 
 }
